requirements.txt
# Core (prototype dependencies — unchanged)
fastapi
uvicorn[standard]
duckdb
pandas
requests
prometheus_client
opentelemetry-api
opentelemetry-sdk
numpy

# Production: async HTTP + connection pooling for connectors
aiohttp>=3.9.0

# Production: Redis distributed cache + distributed rate limiter
redis[hiredis]>=5.0.0

# Production: fast binary serialization for cache values (~30% smaller than JSON)
msgpack>=1.0.7

# Production: SQL AST parsing (replaces regex string matching in engine.py)
sqlglot>=23.0.0

# Production: OIDC/JWT validation (RS256/ES256)
python-jose[cryptography]>=3.3.0

# Production: async HTTP client for OPA + JWKS endpoint calls
httpx>=0.27.0

# Production: TenantConfig Pydantic v2 validation
pydantic>=2.0.0

# Production: tenant YAML config parsing
pyyaml>=6.0.1

# Observability: OpenTelemetry trace export (OTLP + console)
opentelemetry-exporter-otlp-proto-http>=1.20.0
